abstracts to a reusable function
